Move qrypto store logic
