+++ src/components/QryptoPrompt/index.js
@@ -19,12 +19,12 @@ const Message = withStyles(styles)(({ classes, textAlign }) => (
   </Typography>
 ));
 
-const InstallNowButton = inject('store')(({ store: { components: { qryptoPrompt } } }) => (
+const InstallNowButton = inject('store')(({ store: { qrypto } }) => (
   <Button
     variant="raised"
     size="small"
     color="primary"
-    onClick={qryptoPrompt.onInstallClick}
+    onClick={qrypto.onInstallClick}
   >
     <FormattedMessage id="qrypto.installNow" defaultMessage="Install Now" />
   </Button>
@@ -42,9 +42,9 @@ export const InstallQryptoInline = withStyles(styles)(({ classes }) => (
   </div>
 ));
 
-export const InstallQryptoPopover = withStyles(styles)(inject('store')(observer(({ classes, store: { components: { qryptoPrompt } } }) => (
+export const InstallQryptoPopover = withStyles(styles)(inject('store')(observer(({ classes, store: { qrypto } }) => (
   <Popover
-    open={qryptoPrompt.popoverOpen}
+    open={qrypto.popoverOpen}
     anchorOrigin={{ horizontal: 'center', vertical: 'top' }}
     transformOrigin={{ horizontal: 'center', vertical: 'center' }}
   >
@@ -59,7 +59,7 @@ export const InstallQryptoPopover = withStyles(styles)(inject('store')(observer(
             size="small"
             color="primary"
             className={classes.remindButton}
-            onClick={() => qryptoPrompt.popoverOpen = false}
+            onClick={() => qrypto.popoverOpen = false}
           >
             <FormattedMessage id="qrypto.remindMeLater" defaultMessage="Remind Me Later" />
           </Button>
