--- src/components/QryptoPrompt/store.js
@@ -1,18 +1,0 @@
-import { action, observable } from 'mobx';
-
-import { urls } from '../../config/app';
-
-export default class QryptoPromptStore {
-  @observable popoverOpen = false;
-
-  @action
-  open = () => {
-    this.popoverOpen = true;
-  }
-
-  @action
-  onInstallClick = () => {
-    window.open(urls.qryptoWebStore, '_blank');
-    this.popoverOpen = false;
-  }
-}
